requirements.txt
Django==4.0.4
gunicorn==19.9.0
psycopg2==2.9.9
dj_database_url
django-jazzmin==2.6.0
django-cors-headers==4.1.0
Pillow==9.5.0
drf-spectacular==0.27.1
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
django-filter==23.2